Move `TimePeriod` to `types.py` in dbt/artifacts
